Setting icon font weight to Medium
It seems the default font weight is Normal or 400. Changing it to Medium (500) seem to make the text more visible. The rendered text looks less thin and less gray, and instead looks more black without looking too bold. Setting it to DemiBold(600) looks the same as Medium. Setting it to Bold (700) is too much.

https://doc.qt.io/qtforpython-6/PySide6/QtGui/QFont.html#PySide6.QtGui.PySide6.QtGui.QFont.Weight

https://github.com/digitaltrails/vdu_controls/issues/55#issuecomment-1733889397